+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 **Release Highlights:**
 
+- New icon thanks to Flubs! [I HOPE]
+- Add a tweak to Wakes' Splash Plane effect to make them glow just like the wakes [WIP]
 - Switch config to Forge Config API Port
   - The way I eventually want to set up the visual config wouldn't have been possible using YACL, the Forge API port was the best option I could find for an existing config API
 - Enabled applicable fixes on NeoForge >=26.2 [WIP]